Removed type redefinitions.

+++ zlib/lz.c
@@ -3,9 +3,9 @@
 
 #define FIXED_BLOCK_SIZE 65535
 
-static const u8 CL_ORDER[19] = { 16, 17, 18, 0, 8, 7, 9, 6, 10, 5, 11, 4, 12, 3, 13, 2, 14, 1, 15 };
+static const uint8_t CL_ORDER[19] = { 16, 17, 18, 0, 8, 7, 9, 6, 10, 5, 11, 4, 12, 3, 13, 2, 14, 1, 15 };
 
-static const u32 BASE_LENGTH[] = {
+static const uint32_t BASE_LENGTH[] = {
 	3, 4, 5, 6, 7, 8, 9, 10, //257 - 264
 	11, 13, 15, 17,          //265 - 268
 	19, 23, 27, 31,          //269 - 273 
@@ -15,7 +15,7 @@ static const u32 BASE_LENGTH[] = {
 	258                      //285
 };
 
-static const u8 BASE_LENGTH_EXTRA_BITS[] = {
+static const uint8_t BASE_LENGTH_EXTRA_BITS[] = {
 	0, 0, 0, 0, 0, 0, 0, 0, //257 - 264
 	1, 1, 1, 1, //265 - 268
 	2, 2, 2, 2, //269 - 273 
@@ -25,7 +25,7 @@ static const u8 BASE_LENGTH_EXTRA_BITS[] = {
 	0           //285
 };
 
-static const u32 BASE_DIST[] = {
+static const uint32_t BASE_DIST[] = {
 	/*0*/ 1, 2, 3, 4,    //0-3
 	/*1*/ 5, 7,          //4-5
 	/*2*/ 9, 13,         //6-7
@@ -43,7 +43,7 @@ static const u32 BASE_DIST[] = {
 	0   , 0
 };
 
-static const u8 BASE_DIST_EXTRA_BITS[] = {
+static const uint8_t BASE_DIST_EXTRA_BITS[] = {
 	/*0*/ 0, 0, 0, 0, //0-3
 	/*1*/ 1, 1,       //4-5
 	/*2*/ 2, 2,       //6-7
@@ -73,26 +73,26 @@ static const uint8_t FIXED_DIST_CL[] = {
 };
 
 typedef struct HuffEntry {
-	u32 code;
-	u32 length;
+	uint32_t code;
+	uint32_t length;
 } HuffEntry;
 
-#define _HuffTable(n) struct { u64 size; HuffEntry tree[n]; }
+#define _HuffTable(n) struct { uint64_t size; HuffEntry tree[n]; }
 
 typedef struct HuffTable {
-	u64 size;
+	uint64_t size;
 	HuffEntry tree[];
 } HuffTable;
 
-static u32 lzComputeAdler32(const byte* data, u64 size) {
-	u16 s1 = 1;
-	u16 s2 = 0;
+static uint32_t lzComputeAdler32(const uint8_t* data, uint64_t size) {
+	uint32_t s1 = 1;
+	uint32_t s2 = 0;
 
-	for (u64 i = 0; i < size; ++i) {
+	for (uint64_t i = 0; i < size; ++i) {
 		s1 = (s1 + data[i]) % 65521;
 		s2 = (s2 + s1) % 65521;
 	}
-	u32 adler32 = (s2 * 65536) + s1;
+	uint32_t adler32 = (s2 * 65536) + s1;
 
 	return 
 		((adler32 >> 24) & 0x000000FF) |
@@ -101,40 +101,40 @@ static u32 lzComputeAdler32(const byte* data, u64 size) {
 		((adler32 << 24) & 0xFF000000);
 }
 
-static u32 ReverseBits(u32 code, u32 len) {
-	u32 res = 0;
-	for (u32 i = 0; i < len; ++i) {
+static uint32_t ReverseBits(uint32_t code, uint32_t len) {
+	uint32_t res = 0;
+	for (uint32_t i = 0; i < len; ++i) {
 		res = (res << 1) | (code & 1);
 		code >>= 1;
 	}
 	return res;
 }
 
-static void hBuild(HuffTable* table, const u8* codeLens) {
-	u32 BL_COUNT[19] = { 0 };
-	u32 BL_NEXT[19] = { 0 };
+static void hBuild(HuffTable* table, const uint8_t* codeLens) {
+	uint32_t BL_COUNT[19] = { 0 };
+	uint32_t BL_NEXT[19] = { 0 };
 
-	u64 size = table->size;
+	uint64_t size = table->size;
 
-	u32 maxLen = 0;
-	for (u32 i = 0; i < size; ++i) {
-		u32 len = codeLens[i];
+	uint32_t maxLen = 0;
+	for (uint32_t i = 0; i < size; ++i) {
+		uint32_t len = codeLens[i];
 		maxLen = maxLen < len ? len : maxLen;
 	}
 
-	for (u32 i = 0; i < size; ++i) {
+	for (uint32_t i = 0; i < size; ++i) {
 		BL_COUNT[codeLens[i]]++;
 	}
 	BL_COUNT[0] = 0;
 
-	u32 code = 0;
-	for (u32 i = 1; i <= maxLen; ++i) {
+	uint32_t code = 0;
+	for (uint32_t i = 1; i <= maxLen; ++i) {
 		code = (code + BL_COUNT[i - 1]) << 1;
 		BL_NEXT[i] = code;
 	}
 
-	for (u32 i = 0; i <= size; ++i) {
-		u32 len = codeLens[i];
+	for (uint32_t i = 0; i <= size; ++i) {
+		uint32_t len = codeLens[i];
 		if (len != 0) {
 			table->tree[i].length = len;
 			table->tree[i].code = ReverseBits(BL_NEXT[len]++, len);
@@ -146,16 +146,16 @@ static int hDecode(HuffTable* table, BitReader* bs) {
 	DEBUG(table);
 	DEBUG(bs);
 
-	i32 error = 0;
-	u64 size = table->size;
+	int32_t error = 0;
+	uint64_t size = table->size;
 	HuffEntry* tree = table->tree;
 
 	for (int i = 0; i < size; ++i) {
-		u32 len = tree[i].length;
+		uint32_t len = tree[i].length;
 		if (!len) {
 			continue;
 		}
-		u32 code = 0;
+		uint32_t code = 0;
 		CHECK(code = bsPeakBits(bs, len), error);
 		if (code == tree[i].code) {
 			bs->bitBuff >>= len;
@@ -166,11 +166,11 @@ static int hDecode(HuffTable* table, BitReader* bs) {
 	return 1;
 }
 
-int lzInflateInit(ZlibReader* z, const byte* data, u64 size) {
+int lzInflateInit(ZlibReader* z, const uint8_t* data, uint64_t size) {
 	DEBUG(z);
 	DEBUG(data);
 
-	i32 error = 0;
+	int32_t error = 0;
 	CHECK(bsReaderInit(&z->stream, data, size), error);
 	CHECK(z->header.cm = bsGetBits(&z->stream, 4), error);
 	CHECK(z->header.cinfo = bsGetBits(&z->stream, 4), error);
@@ -188,7 +188,7 @@ int lzInflateInit(ZlibReader* z, const byte* data, u64 size) {
 	return 1;
 }
 
-int lzInflate(ZlibReader* z, byte* output, u64 outSize)
+int lzInflate(ZlibReader* z, uint8_t* output, uint64_t outSize)
 {
 	DEBUG(z);
 	DEBUG(output);
@@ -196,13 +196,13 @@ int lzInflate(ZlibReader* z, byte* output, u64 outSize)
 	DEBUG(z->stream.bits);
 
 #ifdef _DEBUG
-	i32 error = 0;
+	int32_t error = 0;
 #endif // _DEBUG
 
-	u64 outIdx = 0;
+	uint64_t outIdx = 0;
 
-	i32 BFINAL = 0;
-	i32 BTYPE = 0;
+	int32_t BFINAL = 0;
+	int32_t BTYPE = 0;
 
 	do
 	{
@@ -214,16 +214,16 @@ int lzInflate(ZlibReader* z, byte* output, u64 outSize)
 		{
 			bsReaderFlush(&z->stream);
 
-			u16 LEN = 0;
-			u16 NLEN = 0;
-			CHECK(LEN = (u16)bsGetBits(&z->stream, 16), error);
-			CHECK(NLEN = (u16)bsGetBits(&z->stream, 16), error);
+			uint16_t LEN = 0;
+			uint16_t NLEN = 0;
+			CHECK(LEN = (uint16_t)bsGetBits(&z->stream, 16), error);
+			CHECK(NLEN = (uint16_t)bsGetBits(&z->stream, 16), error);
 
-			if ((u16)(~LEN) != (u16)(NLEN)) {
+			if ((uint16_t)(~LEN) != (uint16_t)(NLEN)) {
 				return -4;
 			}
 
-			for (u16 i = 0; i < LEN; ++i) {
+			for (uint16_t i = 0; i < LEN; ++i) {
 				if (outIdx > outSize) {
 					return -3;
 				}
@@ -242,7 +242,7 @@ int lzInflate(ZlibReader* z, byte* output, u64 outSize)
 			hBuild((HuffTable*)&dist, FIXED_DIST_CL);
 
 			while (1) {
-				u32 code = 0;
+				uint32_t code = 0;
 				CHECK(code = hDecode((HuffTable*)&litlen, &z->stream), error);
 				if (code == 256)
 					break;
@@ -253,13 +253,13 @@ int lzInflate(ZlibReader* z, byte* output, u64 outSize)
 					output[outIdx++] = code;
 				}
 				else if (code > 256) {
-					u8 baseIdx = code - 257;
-					u32 baseLen = BASE_LENGTH[baseIdx] + bsGetBits(&z->stream, BASE_LENGTH_EXTRA_BITS[baseIdx]);
+					uint8_t baseIdx = code - 257;
+					uint32_t baseLen = BASE_LENGTH[baseIdx] + bsGetBits(&z->stream, BASE_LENGTH_EXTRA_BITS[baseIdx]);
 
-					u8 distanceIdx = hDecode((HuffTable*)&dist, &z->stream);
-					u32 distanceLen = BASE_DIST[distanceIdx] + bsGetBits(&z->stream, BASE_DIST_EXTRA_BITS[distanceIdx]);
+					uint8_t distanceIdx = hDecode((HuffTable*)&dist, &z->stream);
+					uint32_t distanceLen = BASE_DIST[distanceIdx] + bsGetBits(&z->stream, BASE_DIST_EXTRA_BITS[distanceIdx]);
 
-					u64 backIdx = outIdx - distanceLen;
+					uint64_t backIdx = outIdx - distanceLen;
 					while (baseLen--) {
 						if (outIdx > outSize) {
 							return -3;
@@ -272,20 +272,20 @@ int lzInflate(ZlibReader* z, byte* output, u64 outSize)
 		};
 		case 2:
 		{
-			u8 CL[19] = { 0 };
-			u8 HUFF_TREE[320] = { 0 };
+			uint8_t CL[19] = { 0 };
+			uint8_t HUFF_TREE[320] = { 0 };
 
-			u32 HLIT = 0;
-			u32 HDIST = 0;
-			u32 HCLEN = 0;
+			uint32_t HLIT = 0;
+			uint32_t HDIST = 0;
+			uint32_t HCLEN = 0;
 
 			_HuffTable(19) clcl = { 0 };
 			_HuffTable(286) litlen = { 0 };
 			_HuffTable(32) dist = { 0 };
 
-			CHECK((u32)HLIT = (bsGetBits(&z->stream, 5) + 257), error);
-			CHECK((u32)HDIST = (bsGetBits(&z->stream, 5) + 1), error);
-			CHECK((u32)HCLEN = (bsGetBits(&z->stream, 4) + 4), error);
+			CHECK((uint32_t)HLIT = (bsGetBits(&z->stream, 5) + 257), error);
+			CHECK((uint32_t)HDIST = (bsGetBits(&z->stream, 5) + 1), error);
+			CHECK((uint32_t)HCLEN = (bsGetBits(&z->stream, 4) + 4), error);
 
 			if (HLIT > 286 || HDIST > 32 || HCLEN > 19) {
 				return -4;
@@ -295,33 +295,33 @@ int lzInflate(ZlibReader* z, byte* output, u64 outSize)
 			dist.size = HDIST;
 			clcl.size = 19;
 
-			for (u32 i = 0; i < HCLEN; ++i) {
-				CHECK(CL[CL_ORDER[i]] = (u8)bsGetBits(&z->stream, 3), error);
+			for (uint32_t i = 0; i < HCLEN; ++i) {
+				CHECK(CL[CL_ORDER[i]] = (uint8_t)bsGetBits(&z->stream, 3), error);
 			}
 
 			hBuild((HuffTable*)&clcl, CL);
 
-			for (u32 i = 0; i < litlen.size + dist.size;) {
-				u32 code = 0;
+			for (uint32_t i = 0; i < litlen.size + dist.size;) {
+				uint32_t code = 0;
 				CHECK(code = hDecode((HuffTable*)&clcl, &z->stream), error);
 				if (code <= 15) {
 					HUFF_TREE[i++] = code;
 				}
 				else if (code == 16) {
 					DEBUG(i > 0);
-					u32 rep = 0;
-					CHECK(rep = (u32)bsGetBits(&z->stream, 2) + 3, error);
-					for (u32 r = 0; r < rep; r++) HUFF_TREE[i++] = HUFF_TREE[i - 1];
+					uint32_t rep = 0;
+					CHECK(rep = (uint32_t)bsGetBits(&z->stream, 2) + 3, error);
+					for (uint32_t r = 0; r < rep; r++) HUFF_TREE[i++] = HUFF_TREE[i - 1];
 				}
 				else if (code == 17) {
-					u32 rep = 0;
-					CHECK(rep = (u32)bsGetBits(&z->stream, 3) + 3, error);
-					for (u32 r = 0; r < rep; r++) HUFF_TREE[i++] = 0;
+					uint32_t rep = 0;
+					CHECK(rep = (uint32_t)bsGetBits(&z->stream, 3) + 3, error);
+					for (uint32_t r = 0; r < rep; r++) HUFF_TREE[i++] = 0;
 				}
 				else {
-					u32 rep = 0;
-					CHECK(rep = (u32)bsGetBits(&z->stream, 7) + 11, error);
-					for (u32 r = 0; r < rep; r++) HUFF_TREE[i++] = 0;
+					uint32_t rep = 0;
+					CHECK(rep = (uint32_t)bsGetBits(&z->stream, 7) + 11, error);
+					for (uint32_t r = 0; r < rep; r++) HUFF_TREE[i++] = 0;
 				}
 			}
 
@@ -329,7 +329,7 @@ int lzInflate(ZlibReader* z, byte* output, u64 outSize)
 			hBuild((HuffTable*)&dist, HUFF_TREE + litlen.size);
 
 			while (1) {
-				u32 code = 0;
+				uint32_t code = 0;
 				CHECK(code = hDecode((HuffTable*)&litlen, &z->stream), error);
 				if (code == 256) {
 					break;
@@ -341,15 +341,15 @@ int lzInflate(ZlibReader* z, byte* output, u64 outSize)
 					output[outIdx++] = code;
 				}
 				else if (code > 256 && code < 286) {
-					u8 baseIdx = code - 257;
-					u32 baseLen = 0;
+					uint8_t baseIdx = code - 257;
+					uint32_t baseLen = 0;
 					baseLen = BASE_LENGTH[baseIdx] + bsGetBits(&z->stream, BASE_LENGTH_EXTRA_BITS[baseIdx]);
 
-					u8 distIdx = 0;
+					uint8_t distIdx = 0;
 					CHECK(distIdx = hDecode((HuffTable*)&dist, &z->stream), error);
-					u32 basedist = BASE_DIST[distIdx] + bsGetBits(&z->stream, BASE_DIST_EXTRA_BITS[distIdx]);
+					uint32_t basedist = BASE_DIST[distIdx] + bsGetBits(&z->stream, BASE_DIST_EXTRA_BITS[distIdx]);
 
-					u64 backIdx = outIdx - basedist;
+					uint64_t backIdx = outIdx - basedist;
 					while (baseLen--) {
 						if (outIdx > outSize) {
 							return -3;
@@ -369,10 +369,10 @@ int lzInflate(ZlibReader* z, byte* output, u64 outSize)
 
 	} while (BFINAL == 0);
 
-	((byte*)&z->adler32)[0] = bsGetByte(&z->stream);
-	((byte*)&z->adler32)[1] = bsGetByte(&z->stream);
-	((byte*)&z->adler32)[2] = bsGetByte(&z->stream);
-	((byte*)&z->adler32)[3] = bsGetByte(&z->stream);
+	((uint8_t*)&z->adler32)[0] = bsGetByte(&z->stream);
+	((uint8_t*)&z->adler32)[1] = bsGetByte(&z->stream);
+	((uint8_t*)&z->adler32)[2] = bsGetByte(&z->stream);
+	((uint8_t*)&z->adler32)[3] = bsGetByte(&z->stream);
 
 	if (z->adler32 != lzComputeAdler32(output, outIdx)) {
 		return -5;
@@ -381,7 +381,7 @@ int lzInflate(ZlibReader* z, byte* output, u64 outSize)
 	return 1;
 }
 
-int lzDeflateInit(ZlibWriter* z, const byte* data, u64 size)
+int lzDeflateInit(ZlibWriter* z, const uint8_t* data, uint64_t size)
 {
 	DEBUG(z);
 	DEBUG(data);
@@ -393,17 +393,17 @@ int lzDeflateInit(ZlibWriter* z, const byte* data, u64 size)
 	return 1;
 }
 
-int lzDeflate(ZlibWriter* z, byte* output, u64 outSize)
+int lzDeflate(ZlibWriter* z, uint8_t* output, uint64_t outSize)
 {
 	DEBUG(z);
 	DEBUG(output);
 	DEBUG(outSize);
 #ifdef _DEBUG
-	i32 error = 0;
+	int32_t error = 0;
 #endif // _DEBUG
 
-	u64 outIdx = 0;
-	u64 remaining = z->size;
+	uint64_t outIdx = 0;
+	uint64_t remaining = z->size;
 	CHECK(bsWriterInit(&z->stream, output, outSize), error);
 
 	z->header.cm = 0x08;
@@ -412,9 +412,9 @@ int lzDeflate(ZlibWriter* z, byte* output, u64 outSize)
 	z->header.fdict = 0x00;
 	z->header.flevel = 0x00;
 
-	CHECK(bsWriteBits(&z->stream, *(u64*)&z->header, 16), error);
+	CHECK(bsWriteBits(&z->stream, *(uint64_t*)&z->header, 16), error);
 
-	u8 BFINAL = 0;
+	uint8_t BFINAL = 0;
 	do {
 		BFINAL = z->size < FIXED_BLOCK_SIZE;
 
@@ -422,11 +422,11 @@ int lzDeflate(ZlibWriter* z, byte* output, u64 outSize)
 		block.BFINAL = BFINAL;
 		block.BTYPE = 0x00;
 
-		CHECK(bsWriteBits(&z->stream, *(u64*)&block, 3), error);
+		CHECK(bsWriteBits(&z->stream, *(uint64_t*)&block, 3), error);
 		bsWriterFlush(&z->stream);
 
-		u16 LEN = (u16)BFINAL ? (u16)z->size : FIXED_BLOCK_SIZE;
-		u16 NLEN = ~LEN;
+		uint16_t LEN = (uint16_t)BFINAL ? (uint16_t)z->size : FIXED_BLOCK_SIZE;
+		uint16_t NLEN = ~LEN;
 		remaining -= LEN;
 
 		CHECK(bsWriteBits(&z->stream, LEN, 16), error);
@@ -437,7 +437,7 @@ int lzDeflate(ZlibWriter* z, byte* output, u64 outSize)
 
 	} while (!BFINAL);
 
-	u32 adler32 = lzComputeAdler32(z->uncompressed, z->size);
+	uint32_t adler32 = lzComputeAdler32(z->uncompressed, z->size);
 
 	CHECK(bsWriteBits(&z->stream, adler32, 32), error);
